requirements.txt
google-cloud-storage
pdfminer.six
openai==0.10.2
google-api-python-client
google-cloud
google-cloud-bigquery
google-api-python-client
sqlalchemy
langchain
langchain_google_vertexai
google-cloud-bigquery-storage
sqlalchemy-bigquery
google-cloud-pubsub
lxml
requests
pdfkit
Flask
gunicorn
werkzeug==2.2.2
weasyprint
PyPDF2
faiss-cpu
tiktoken
langchain_community
google-cloud-aiplatform
flask_cors